functions/recognizer/recognizer__extract/requirements.txt
fastapi
uvicorn
opencv-python-headless
Pillow
numpy
redis
pytesseract
boto3
